Put common Go code into packages under pkg/...
The `tracer` util in tests and users of tcptracer-bpf (e.g. scope) will
share the "guessing code", hence put it into a separate package to allow
re-use.
